Nitrite native support #1298
