+++ docs/modules/ROOT/partials/reference/components/nitrite.adoc
@@ -2,11 +2,11 @@
 // This file was generated by camel-quarkus-maven-plugin:update-extension-doc-page
 :cq-artifact-id: camel-quarkus-nitrite
 :cq-artifact-id-base: nitrite
-:cq-native-supported: false
-:cq-status: Preview
+:cq-native-supported: true
+:cq-status: Stable
 :cq-deprecated: false
 :cq-jvm-since: 1.0.0
-:cq-native-since: n/a
+:cq-native-since: 1.8.0
 :cq-camel-part-name: nitrite
 :cq-camel-part-title: Nitrite
 :cq-camel-part-description: Access Nitrite databases.
