+++ extensions-jvm/pom.xml
@@ -92,7 +92,6 @@
         <module>mllp</module>
         <module>mvel</module>
         <module>mybatis</module>
-        <module>nitrite</module>
         <module>ognl</module>
         <module>openstack</module>
         <module>printer</module>
