+++ extensions/nitrite/runtime/src/main/resources/META-INF/native-image/serialization-config.json
@@ -1,0 +1,41 @@
+[
+  {
+    "name" : "org.dizitart.no2.NitriteId"
+  },
+  {
+    "name" : "java.lang.Long"
+  },
+  {
+    "name" : "java.lang.Number"
+  },
+  {
+    "name" : "org.dizitart.no2.Document"
+  },
+  {
+    "name" : "java.util.LinkedHashMap"
+  },
+  {
+    "name" : "java.util.HashMap"
+  },
+  {
+    "name" : "java.util.Date"
+  },
+  {
+    "name" : "java.lang.String"
+  },
+  {
+    "name" : "java.lang.Integer"
+  },
+  {
+    "name" : "org.dizitart.no2.meta.Attributes"
+  },
+  {
+    "name" : "org.dizitart.no2.internals.IndexMetaService$IndexMeta"
+  },
+  {
+    "name" : "java.util.concurrent.atomic.AtomicBoolean"
+  },
+  {
+    "name" : "org.dizitart.no2.Index"
+  }
+]
