+++ extensions/pom.xml
@@ -181,6 +181,7 @@
         <module>nats</module>
         <module>netty</module>
         <module>netty-http</module>
+        <module>nitrite</module>
         <module>nsq</module>
         <module>oaipmh</module>
         <module>olingo4</module>
