+++ integration-tests/nitrite/src/main/resources/my-serialization-config.json
@@ -1,0 +1,7 @@
+[
+  {
+    "name" : "org.apache.camel.quarkus.component.nitrite.it.EmployeeSerializable"
+  },{
+    "name" : "org.apache.camel.quarkus.component.nitrite.it.EmployeeMappable"
+  }
+]
