+++ integration-tests/nitrite/src/test/java/org/apache/camel/quarkus/component/nitrite/it/NitriteTestResource.java
