+++ tooling/scripts/test-categories.yaml
@@ -66,6 +66,7 @@ database:
   - leveldb
   - minio
   - mongodb
+  - nitrite
   - pgevent
   - pg-replication-slot
   - sql
